moood_mobile_auth/requirements.txt
fastapi
uvicorn
JPype1
